Refine shared query boundaries
